Do not change to log directory as sudoer, strip parent path instead
Changing to the log directory fails when the suoder user does not have
permissions to explore the directory under its normal privileges.

I have modified the script to search the directory directly,
using its path, while changing to the directory as root user
part of the tar command, while removing the prefix from the archive.