Handle files in subfolders of the given log folder
I had zealously stripped the whole parent path previously,
when only a prefix limited to the given log folder shall be removed.

The prefix is computed as '${fromDirectory%%/}/' to strip any trailing
slash from the directory, if provided by the user, before adding a
trailing slash to get a relative path, and not an absolute path (starting
with /) for the file selected for addition to the archive.