Test archive file size instead of tar result code
tar returns a non-zero code when a file changes while reading,
while the resulting archive is fine, and even contains a copy
of the changed file. So everything is fine, the warning is useful
as information but should not prevent completion of the export.

I now test the size of the archive created by tar, and if non-zero,
considers that it should be exported.